+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 2.x branch
 ## 2.5 banch
+### 2.5.2
+* fixed shell output style;
+* removed useless methods.
+
 ### 2.5.1
 * fixed little bugs.